move building stuff to its own module
