+++ game/magic/city/city.go
@@ -8,180 +8,8 @@ import (
     "github.com/kazzmir/master-of-magic/lib/fraction"
     "github.com/kazzmir/master-of-magic/game/magic/data"
     "github.com/kazzmir/master-of-magic/game/magic/units"
+    "github.com/kazzmir/master-of-magic/game/magic/building"
 )
-
-type Building int
-const (
-    BuildingNone Building = iota
-
-    BuildingBarracks
-    BuildingArmory
-    BuildingFightersGuild
-    BuildingArmorersGuild
-    BuildingWarCollege
-    BuildingSmithy
-    BuildingStables
-    BuildingAnimistsGuild
-    BuildingFantasticStable
-    BuildingShipwrightsGuild
-    BuildingShipYard
-    BuildingMaritimeGuild
-    BuildingSawmill
-    BuildingLibrary
-    BuildingSagesGuild
-    BuildingOracle
-    BuildingAlchemistsGuild
-    BuildingUniversity
-    BuildingWizardsGuild
-    BuildingShrine
-    BuildingTemple
-    BuildingParthenon
-    BuildingCathedral
-    BuildingMarketplace
-    BuildingBank
-    BuildingMerchantsGuild
-    BuildingGranary
-    BuildingFarmersMarket
-    BuildingForestersGuild
-    BuildingBuildersHall
-    BuildingMechaniciansGuild
-    BuildingMinersGuild
-    BuildingCityWalls
-    BuildingFortress
-    BuildingSummoningCircle
-
-    BuildingHousing
-    BuildingTradeGoods
-
-    // not a real building, just a marker
-    BuildingLast
-)
-
-// FIXME: these values can come from buildat.lbx
-func (building Building) ProductionCost() int {
-    switch building {
-        case BuildingBarracks: return 30
-        case BuildingArmory: return 80
-        case BuildingFightersGuild: return 200
-        case BuildingArmorersGuild: return 350
-        case BuildingWarCollege: return 500
-        case BuildingSmithy: return 40
-        case BuildingStables: return 80
-        case BuildingAnimistsGuild: return 300
-        case BuildingFantasticStable: return 600
-        case BuildingShipwrightsGuild: return 100
-        case BuildingShipYard: return 200
-        case BuildingMaritimeGuild: return 400
-        case BuildingSawmill: return 100
-        case BuildingLibrary: return 60
-        case BuildingSagesGuild: return 120
-        case BuildingOracle: return 500
-        case BuildingAlchemistsGuild: return 250
-        case BuildingUniversity: return 300
-        case BuildingWizardsGuild: return 1000
-        case BuildingShrine: return 100
-        case BuildingTemple: return 200
-        case BuildingParthenon: return 400
-        case BuildingCathedral: return 800
-        case BuildingMarketplace: return 100
-        case BuildingBank: return 250
-        case BuildingMerchantsGuild: return 600
-        case BuildingGranary: return 40
-        case BuildingFarmersMarket: return 100
-        case BuildingForestersGuild: return 200
-        case BuildingBuildersHall: return 60
-        case BuildingMechaniciansGuild: return 600
-        case BuildingMinersGuild: return 300
-        case BuildingCityWalls: return 150
-    }
-
-    return 0
-}
-
-func (building Building) UpkeepCost() int {
-    switch building {
-        case BuildingBarracks: return 0
-        case BuildingArmory: return 2
-        case BuildingFightersGuild: return 3
-        case BuildingArmorersGuild: return 4
-        case BuildingWarCollege: return 5
-        case BuildingSmithy: return 1
-        case BuildingStables: return 2
-        case BuildingAnimistsGuild: return 5
-        case BuildingFantasticStable: return 6
-        case BuildingShipwrightsGuild: return 1
-        case BuildingShipYard: return 2
-        case BuildingMaritimeGuild: return 4
-        case BuildingSawmill: return 2
-        case BuildingLibrary: return 1
-        case BuildingSagesGuild: return 2
-        case BuildingOracle: return 4
-        case BuildingAlchemistsGuild: return 3
-        case BuildingUniversity: return 3
-        case BuildingWizardsGuild: return 5
-        case BuildingShrine: return 1
-        case BuildingTemple: return 2
-        case BuildingParthenon: return 3
-        case BuildingCathedral: return 4
-        case BuildingMarketplace: return 1
-        case BuildingBank: return 3
-        case BuildingMerchantsGuild: return 5
-        case BuildingGranary: return 1
-        case BuildingFarmersMarket: return 2
-        case BuildingForestersGuild: return 2
-        case BuildingBuildersHall: return 1
-        case BuildingMechaniciansGuild: return 5
-        case BuildingMinersGuild: return 3
-        case BuildingCityWalls: return 2
-    }
-
-    return 0
-}
-
-func (building Building) String() string {
-    switch building {
-        case BuildingBarracks: return "Barracks"
-        case BuildingArmory: return "Armory"
-        case BuildingFightersGuild: return "Fighters Guild"
-        case BuildingArmorersGuild: return "Armorers Guild"
-        case BuildingWarCollege: return "War College"
-        case BuildingSmithy: return "Smithy"
-        case BuildingStables: return "Stables"
-        case BuildingAnimistsGuild: return "Animist's Guild"
-        case BuildingFantasticStable: return "Fantastic Stable"
-        case BuildingShipwrightsGuild: return "Ship Wrights Guild"
-        case BuildingShipYard: return "Shipyard"
-        case BuildingMaritimeGuild: return "Maritime Guild"
-        case BuildingSawmill: return "Sawmill"
-        case BuildingLibrary: return "Library"
-        case BuildingSagesGuild: return "Sage's Guild"
-        case BuildingOracle: return "Oracle"
-        case BuildingAlchemistsGuild: return "Alchemists Guild"
-        case BuildingUniversity: return "University"
-        case BuildingWizardsGuild: return "Wizard's Guild"
-        case BuildingShrine: return "Shrine"
-        case BuildingTemple: return "Temple"
-        case BuildingParthenon: return "Parthenon"
-        case BuildingCathedral: return "Cathedral"
-        case BuildingMarketplace: return "Marketplace"
-        case BuildingBank: return "Bank"
-        case BuildingMerchantsGuild: return "Merchant's Guild"
-        case BuildingGranary: return "Granary"
-        case BuildingFarmersMarket: return "Farmer's Market"
-        case BuildingForestersGuild: return "Forester's Guild"
-        case BuildingBuildersHall: return "Builder's Hall"
-        case BuildingMechaniciansGuild: return "Mechanician's Guild"
-        case BuildingMinersGuild: return "Miner's Guild"
-        case BuildingCityWalls: return "City Walls"
-        case BuildingFortress: return "Fortress"
-        case BuildingSummoningCircle: return "Summoning Circle"
-
-        case BuildingHousing: return "Housing"
-        case BuildingTradeGoods: return "Trade Goods"
-    }
-
-    return "?"
-}
 
 type CityEvent interface {
 }
@@ -235,7 +63,7 @@ type City struct {
     X int
     Y int
     Banner data.BannerType
-    Buildings *set.Set[Building]
+    Buildings *set.Set[building.Building]
 
     TaxRate fraction.Fraction
 
@@ -244,7 +72,7 @@ type City struct {
 
     // how many hammers the city has produced towards the current project
     Production float32
-    ProducingBuilding Building
+    ProducingBuilding building.Building
     ProducingUnit units.Unit
 }
 
@@ -254,7 +82,7 @@ func MakeCity(name string, x int, y int, race data.Race, taxRate fraction.Fracti
         X: x,
         Y: y,
         Race: race,
-        Buildings: set.MakeSet[Building](),
+        Buildings: set.MakeSet[building.Building](),
         TaxRate: taxRate,
     }
 
@@ -266,7 +94,7 @@ func (city *City) UpdateTaxRate(taxRate fraction.Fraction, garrison []*units.Ove
     city.UpdateUnrest(garrison)
 }
 
-func (city *City) AddBuilding(building Building){
+func (city *City) AddBuilding(building building.Building){
     city.Buildings.Insert(building)
 }
 
@@ -430,27 +258,27 @@ func (city *City) ComputeUnrest(garrison []*units.OverworldUnit) int {
     // pacificationRetort = 1.5
 
     pacification := float64(0)
-    if city.Buildings.Contains(BuildingShrine) {
+    if city.Buildings.Contains(building.BuildingShrine) {
         pacification += 1 * pacificationRetort
     }
 
-    if city.Buildings.Contains(BuildingTemple) {
+    if city.Buildings.Contains(building.BuildingTemple) {
         pacification += float64(templePacification(city.Race)) * pacificationRetort
     }
 
-    if city.Buildings.Contains(BuildingParthenon) {
+    if city.Buildings.Contains(building.BuildingParthenon) {
         pacification += float64(parthenonPacification(city.Race)) * pacificationRetort
     }
 
-    if city.Buildings.Contains(BuildingCathedral) {
+    if city.Buildings.Contains(building.BuildingCathedral) {
         pacification += float64(cathedralPaclification(city.Race)) * pacificationRetort
     }
 
-    if city.Buildings.Contains(BuildingAnimistsGuild) {
+    if city.Buildings.Contains(building.BuildingAnimistsGuild) {
         pacification += float64(animistsGuildPacification(city.Race))
     }
 
-    if city.Buildings.Contains(BuildingOracle) {
+    if city.Buildings.Contains(building.BuildingOracle) {
         pacification += float64(oraclePacification(city.Race))
     }
 
@@ -468,11 +296,11 @@ func (city *City) MaximumCitySize() int {
 
     bonus := 0
 
-    if city.Buildings.Contains(BuildingGranary) {
+    if city.Buildings.Contains(building.BuildingGranary) {
         bonus += 1
     }
 
-    if city.Buildings.Contains(BuildingFarmersMarket) {
+    if city.Buildings.Contains(building.BuildingFarmersMarket) {
         bonus += 3
     }
 
@@ -500,11 +328,11 @@ func (city *City) PopulationGrowthRate() int {
         case data.RaceTroll: base -= 20
     }
 
-    if city.Buildings.Contains(BuildingGranary) {
+    if city.Buildings.Contains(building.BuildingGranary) {
         base += 20
     }
 
-    if city.Buildings.Contains(BuildingFarmersMarket) {
+    if city.Buildings.Contains(building.BuildingFarmersMarket) {
         base += 30
     }
 
@@ -541,7 +369,7 @@ func (city *City) foodProductionRate(farmers int) int {
 
     baseRate := float32(rate * farmers)
 
-    if city.Buildings.Contains(BuildingForestersGuild) {
+    if city.Buildings.Contains(building.BuildingForestersGuild) {
         baseRate += 2
     }
 
@@ -554,11 +382,11 @@ func (city *City) foodProductionRate(farmers int) int {
 
     bonus := 0
 
-    if city.Buildings.Contains(BuildingGranary) {
+    if city.Buildings.Contains(building.BuildingGranary) {
         bonus += 2
     }
 
-    if city.Buildings.Contains(BuildingFarmersMarket) {
+    if city.Buildings.Contains(building.BuildingFarmersMarket) {
         bonus += 3
     }
 
@@ -582,7 +410,7 @@ func (city *City) MoneyProductionRate() int {
 
     bonus := float32(0)
 
-    if city.ProducingBuilding == BuildingTradeGoods {
+    if city.ProducingBuilding == building.BuildingTradeGoods {
         bonus = city.WorkProductionRate() / 2
     }
 
@@ -638,7 +466,7 @@ func (city *City) DoNextTurn(garrison []*units.OverworldUnit) []CityEvent {
             if city.Production >= float32(city.ProducingBuilding.ProductionCost()) {
                 city.Buildings.Insert(city.ProducingBuilding)
                 city.Production = 0
-                city.ProducingBuilding = BuildingTradeGoods
+                city.ProducingBuilding = building.BuildingHousing
             }
         } else if !city.ProducingUnit.Equals(units.UnitNone) && city.Production >= float32(city.ProducingUnit.ProductionCost) {
             cityEvents = append(cityEvents, &CityEventNewUnit{Unit: city.ProducingUnit})
